Revert "Reapply "Attempted to improve transfer buffer scan, giving up.""
This reverts commit 42e1bd8a4a410767ce622b3c367179b08a7b8aac.

+++ inc/common.hpp
@@ -171,8 +171,8 @@ struct GPU_Data
     uint64_t* current_level;
     int* current_task;
     // count
-    uint64_t* total_tasks;
-    uint64_t* total_cliques;
+    int* total_tasks;
+    int* total_cliques;
     // EXPANSION HELPERS
     Vertex* temp_vertex_array;
     int* vertex_order_map;
